App/functions/delta-dtm-transsect-plot/pyproject.toml
[project]
name = "delta-dtm-line-plot"
version = "0.1.0"
description = "Add your description here"
# readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.6",
    "azure-storage-blob>=12.24.1",
    "duckdb>=1.1.3",
    "fastapi>=0.115.8",
    "fastparquet>=2024.11.0",
    "geopandas>=1.0.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.0",
    "pyarrow>=19.0.1",
    "pydantic-shapely>=1.0.0a4",
    "pystac-client>=0.8.5",
    "rioxarray>=0.18.2",
    "uvicorn>=0.34.0",
    "xarray>=2024.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.4",
]
